Fixed create_hourly_projects
The for loop inside that was chacking to see if a project was already
present wasn't breaking out when it found the project.  I rewrote it in
a while loop to make it easier to parse.  I also changed around how the
'Targeted Residences' were created, and added a function to delete all
projects.

This is the code as run for the first real hourly list.